IP 1.20.1: Fixed overrendering error where the chart would render for each new input. Now inputs are buffered first and then rendered in the end. Fixed double for weight loss.

+++ gym.backend/onlinefitnessmanager/src/main/java/org/unibl/etf/onlinefitnessmanager/model/entities/CompletedExerciseEntity.java
@@ -27,7 +27,7 @@ public class CompletedExerciseEntity {
     private LocalDate dayOfCompletion;
     @Basic
     @Column(name = "weight_loss", nullable = false)
-    private Integer weightLoss;
+    private Double weightLoss;
     @Basic
     @Column(name = "result_description", nullable = true, length = 64)
     private String resultDescription;
@@ -76,11 +76,11 @@ public class CompletedExerciseEntity {
         this.dayOfCompletion = dayOfCompletion;
     }
 
-    public Integer getWeightLoss() {
+    public Double getWeightLoss() {
         return weightLoss;
     }
 
-    public void setWeightLoss(Integer weightLoss) {
+    public void setWeightLoss(Double weightLoss) {
         this.weightLoss = weightLoss;
     }
